requirements.txt
click==7.1.2
colorama==0.4.4
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pypng==0.0.20
PyQRCode==1.2.1
qrcode==6.1
six==1.15.0
Werkzeug==1.0.1
